Registered the save task as background task
This makes the dialog that pops up if background tasks are running wait
for pending saves.
